[Spa] Introduce NoSpaFrontEnd option for testing purposes
In order for us to ensure the templates are up to date with the
latest changes on the .NET runtime, we want to have some tests
that are able to build them (and run the server part independently)
if necessary.

For that matter, we are introducing the NoSpaFrontEnd option which is
hidden from the users and that can be specified in tests to remove all
integration with the SPA frontend.

We will continue droping the files on disk, but there will be no npm
restore during build nor we will try to launch the SPA proxy when the
app launches.

We will have some tests on the ASP.NET Core repo that install the
templates package produced by the repo and build the templates to
at least catch build breaks that changes on ASP.NET Core can cause.