Make `ensureUserManagerInitialized` reentrant (avoids multiple iframes)
This PR solves an issue where multiple `checksession` iframes are created. This happens because every method in the *Authorize* service first calls `ensureUserManagerInitialized`, and this function first checks if the member `userManager` is truthy before continuing. However this function is not truly reentrant, because of `async` and multiple calls to it can pass over the `userManager` truthy check.

The solution involves wrapping the original function, and making sure the internal function is only called once, by passing the truthy check up from the `userManager` to the promise of initializing it.